refactor(products): move key-value operation utilities to core (#138)
refactor(products): move key-value operation utilities to core
