+++ src/datasources/products/types.ts
@@ -93,8 +93,3 @@ export interface QBField extends QueryBuilderField {
     }>;
   },
 }
-
-export interface PropertyFieldKeyValuePair {
-  key: string;
-  value: string | number;
-};
